refactor(courtly): standardize directory and file case sensitivity
